# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Система бронирования номеров",
        "contact": {}
    },
    "paths": {
...
        "/bookings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Бронирование номера только для авторизованных пользователей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Бронирование номера",
                "parameters": [
                    {
                        "description": "Данные для бронирования",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookings.CreateBookingInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Данные о бранировании",
                        "schema": {
                            "$ref": "#/definitions/response.BookingResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Номер не найден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Номер уже забронирован в этот период",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка при проверке доступности номера или при создании бронирования",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отмена бронирования пользователем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Отмена бронирования",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование успешно отменено",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Бронирование уже оплачено и не может быть отменено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Вы не можете отменить бронирование, которое не принадлежит вам",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Бронирование не найдено",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка при отмене бронирования",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/email/test": {
            "get": {
                "description": "Отправляет тестовое письмо на указанную почту.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Отправка тестового письма",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email получателя",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тестовое письмо успешно отправлено",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Не указан email получателя",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка отправки письма",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/owners/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение бронирований для владельца",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Получение бронирований владельца",
                "responses": {
                    "201": {
                        "description": "Данные о бранировании",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.BookingResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении бронирований",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{id}/bookings": {
            "get": {
                "description": "Получение бронирований для номера",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Получение бронирований для номера",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID номера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Данные о бранировании",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.BookingResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении списка бронирований",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "bookings.CreateBookingInput": {
            "type": "object",
            "required": [
                "end_date",
                "room_id",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "response.BookingResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "payment_status": {
                    "description": "Статус оплаты",
                    "type": "string"
                },
                "room_id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "total_cost": {
                    "description": "Итоговая стоимость",
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.ErrorResponse": {
            "description": "Стандартный ответ при ошибке",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.SuccessResponse": {
            "description": "Стандартный ответ при успешном выполнении",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}